+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -140,7 +140,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-13T08:48:57.409Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-13T10:17:23.749Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")