+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -193,7 +193,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kev8pwdtwq1p10zbe3c820rc"
+        "conv_01kevdrredpyp5vxpe13yymphd"
     )
     r"""Unique conversation identifier with `conv_` prefix."""