+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetitemop.py
@@ -1166,7 +1166,7 @@ class Evaluations3(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsDatasetsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-13T08:49:11.187Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-13T10:17:35.464Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1218,7 +1218,7 @@ class Evaluations2(BaseModel):
 
     source: Optional[CreateDatasetItemEvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-13T08:49:11.186Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-13T10:17:35.464Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1270,7 +1270,7 @@ class Evaluations1(BaseModel):
 
     source: Optional[EvaluationsSource] = "orq"
 
-    reviewed_at: Optional[datetime] = parse_datetime("2026-01-13T08:49:11.186Z")
+    reviewed_at: Optional[datetime] = parse_datetime("2026-01-13T10:17:35.463Z")
     r"""The date and time the item was reviewed"""
 
 
@@ -1349,5 +1349,5 @@ class CreateDatasetItemResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-13T08:48:57.409Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-13T10:17:23.749Z")
     r"""The date and time the resource was last updated"""