+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -130,7 +130,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KEV8PWWTQE19FK31WRS6XX65"
+        "01KEVDRRW4CRY9DYEP2T2EZFHT"
     )
     r"""The unique identifier of the data source"""