+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KEV8PWWZ9ASB1DSVHNC1EJRF"
+        "01KEVDRRWGCS1017RNR3C9Y1T1"
     )
     r"""The unique identifier of the data source"""