+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -193,7 +193,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KEV8PWQXPRESP47178YHA7ZQ"
+    id: Optional[str] = "01KEVDRRQRJV5TKS39411VZ3SK"
 
     description: Optional[str] = None
 
@@ -916,7 +916,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KEV8PWQMFTBN8RJXK6PY09TJ"
+        "tool_01KEVDRRQ846WD9JZZ06A0EQKA"
     )
 
     display_name: Optional[str] = None
@@ -990,7 +990,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KEV8PWQK2K3EVAVGFHQA5D1B"
+    id: Optional[str] = "01KEVDRRQ73G8TVPM2EJAHBSJJ"
 
     description: Optional[str] = None
 
@@ -1089,7 +1089,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KEV8PWQJR5GTRMB776772GJR"
+        "tool_01KEVDRRQ672HXTZBYRP06B5R9"
     )
 
     display_name: Optional[str] = None
@@ -1301,7 +1301,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KEV8PWQEF2EEXX5QC8RKNA50"
+        "tool_01KEVDRRQ3PWDQP92B7ZPY06JP"
     )
 
     display_name: Optional[str] = None
@@ -1455,7 +1455,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KEV8PWQD2282TV438Z43EAQQ"
+        "tool_01KEVDRRPV25B0B5FQ3960JFX2"
     )
 
     display_name: Optional[str] = None
@@ -1613,7 +1613,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KEV8PWQB7MR6653H2GHA29AS"
+        "tool_01KEVDRRPTFRHZ86XV4B7ZBYTY"
     )
 
     display_name: Optional[str] = None