+++ packages/orq-rc/RELEASES.md
@@ -4619,3 +4619,13 @@ Based on:
 - [python v4.2.0-rc.24] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.24] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.24 - packages/orq-rc
+
+## 2026-01-13 10:19:52
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.685.0 (2.794.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.25] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.25 - packages/orq-rc